requirements.txt
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
dominate==2.8.0
Flask==2.1.3
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==4.7.0
Flask-Moment==1.0.4
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.3.0
py==1.11.0
pytest==7.1.3
requests==2.28.1
requests-toolbelt==0.9.1
SQLAlchemy==1.4.40
tomli==2.0.1
urllib3==1.26.16
visitor==0.1.3
Werkzeug==2.3.7
WTForms==3.0.1
load-dotenv==0.1.0